requirements.txt
aenum==3.1.15
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
Deprecated==1.2.14
frozenlist==1.4.1
future==1.0.0
idna==3.7
line-bot-sdk==3.11.0
multidict==6.0.5
netifaces==0.11.0
pydantic==2.7.2
pydantic_core==2.18.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
typing_extensions==4.12.1
urllib3==2.2.1
wrapt==1.16.0
yarl==1.9.4
